# rt: use task::Inject with current_thread scheduler