--- a/tokio/src/runtime/task/inject.rs
+++ b/tokio/src/runtime/task/inject.rs
@@ -65,10 +65,6 @@ impl<T: 'static> Inject<T> {
         true
     }
 
-    pub(crate) fn is_closed(&self) -> bool {
-        self.pointers.lock().is_closed
-    }
-
     pub(crate) fn len(&self) -> usize {
         self.len.load(Acquire)
     }
@@ -104,71 +100,6 @@ impl<T: 'static> Inject<T> {
         self.len.store(len + 1, Release);
     }
 
-    /// Pushes several values into the queue.
-    #[inline]
-    pub(crate) fn push_batch<I>(&self, mut iter: I)
-    where
-        I: Iterator<Item = task::Notified<T>>,
-    {
-        let first = match iter.next() {
-            Some(first) => first.into_raw(),
-            None => return,
-        };
-
-        // Link up all the tasks.
-        let mut prev = first;
-        let mut counter = 1;
-
-        // We are going to be called with an `std::iter::Chain`, and that
-        // iterator overrides `for_each` to something that is easier for the
-        // compiler to optimize than a loop.
-        iter.for_each(|next| {
-            let next = next.into_raw();
-
-            // safety: Holding the Notified for a task guarantees exclusive
-            // access to the `queue_next` field.
-            set_next(prev, Some(next));
-            prev = next;
-            counter += 1;
-        });
-
-        // Now that the tasks are linked together, insert them into the
-        // linked list.
-        self.push_batch_inner(first, prev, counter);
-    }
-
-    /// Inserts several tasks that have been linked together into the queue.
-    ///
-    /// The provided head and tail may be be the same task. In this case, a
-    /// single task is inserted.
-    #[inline]
-    fn push_batch_inner(
-        &self,
-        batch_head: NonNull<task::Header>,
-        batch_tail: NonNull<task::Header>,
-        num: usize,
-    ) {
-        debug_assert!(get_next(batch_tail).is_none());
-
-        let mut p = self.pointers.lock();
-
-        if let Some(tail) = p.tail {
-            set_next(tail, Some(batch_head));
-        } else {
-            p.head = Some(batch_head);
-        }
-
-        p.tail = Some(batch_tail);
-
-        // Increment the count.
-        //
-        // safety: All updates to the len atomic are guarded by the mutex. As
-        // such, a non-atomic load followed by a store is safe.
-        let len = unsafe { self.len.unsync_load() };
-
-        self.len.store(len + num, Release);
-    }
-
     pub(crate) fn pop(&self) -> Option<task::Notified<T>> {
         // Fast path, if len == 0, then there are no values
         if self.is_empty() {
@@ -198,6 +129,79 @@ impl<T: 'static> Inject<T> {
 
         // safety: a `Notified` is pushed into the queue and now it is popped!
         Some(unsafe { task::Notified::from_raw(task) })
+    }
+}
+
+cfg_rt_multi_thread! {
+    impl<T: 'static> Inject<T> {
+        pub(crate) fn is_closed(&self) -> bool {
+            self.pointers.lock().is_closed
+        }
+
+        /// Pushes several values into the queue.
+        #[inline]
+        pub(crate) fn push_batch<I>(&self, mut iter: I)
+        where
+            I: Iterator<Item = task::Notified<T>>,
+        {
+            let first = match iter.next() {
+                Some(first) => first.into_raw(),
+                None => return,
+            };
+
+            // Link up all the tasks.
+            let mut prev = first;
+            let mut counter = 1;
+
+            // We are going to be called with an `std::iter::Chain`, and that
+            // iterator overrides `for_each` to something that is easier for the
+            // compiler to optimize than a loop.
+            iter.for_each(|next| {
+                let next = next.into_raw();
+
+                // safety: Holding the Notified for a task guarantees exclusive
+                // access to the `queue_next` field.
+                set_next(prev, Some(next));
+                prev = next;
+                counter += 1;
+            });
+
+            // Now that the tasks are linked together, insert them into the
+            // linked list.
+            self.push_batch_inner(first, prev, counter);
+        }
+
+        /// Inserts several tasks that have been linked together into the queue.
+        ///
+        /// The provided head and tail may be be the same task. In this case, a
+        /// single task is inserted.
+        #[inline]
+        fn push_batch_inner(
+            &self,
+            batch_head: NonNull<task::Header>,
+            batch_tail: NonNull<task::Header>,
+            num: usize,
+        ) {
+            debug_assert!(get_next(batch_tail).is_none());
+
+            let mut p = self.pointers.lock();
+
+            if let Some(tail) = p.tail {
+                set_next(tail, Some(batch_head));
+            } else {
+                p.head = Some(batch_head);
+            }
+
+            p.tail = Some(batch_tail);
+
+            // Increment the count.
+            //
+            // safety: All updates to the len atomic are guarded by the mutex. As
+            // such, a non-atomic load followed by a store is safe.
+            let len = unsafe { self.len.unsync_load() };
+
+            self.len.store(len + num, Release);
+        }
     }
 }