Depends on #5701

Previously, the current_thread scheduler used its own injection queue instead of sharing the same one as the multi-threaded scheduler. This patch updates the current_thread scheduler to use the same injection queue as the multi-threaded one (`task::Inject`).

`task::Inject` includes an optimization where it does not need to acquire the mutex if the queue is empty.

The primary objective for this PR is to unify the injection queue code paths for future tweaks. The benchmark results between the current `master` and this change are within the margin of error.

**Before:**

```
test spawn_many_local       ... bench:     222,721 ns/iter (+/- 5,288)
test spawn_many_remote_busy ... bench:   5,576,420 ns/iter (+/- 175,418)
test spawn_many_remote_idle ... bench:     228,768 ns/iter (+/- 1,763)
```

**After:**

```
test spawn_many_local       ... bench:     216,111 ns/iter (+/- 4,268)
test spawn_many_remote_busy ... bench:   5,329,780 ns/iter (+/- 143,615)
test spawn_many_remote_idle ... bench:     212,560 ns/iter (+/- 2,962)
````