@hawkw 2023-05-19
It looks like the `close`, `push_batch`, and `push_batch_inner` methods on the injector queue need to be behind `cfg_rt_multi_thread!`, as they're unused when building without `features = "rt-multi-thread"`: https://github.com/tokio-rs/tokio/actions/runs/5026769568/jobs/9015455907?pr=5702#step:7:175