@hawkw 2023-05-19
Looks like [the WASM tests are failing](https://github.com/tokio-rs/tokio/actions/runs/5027141640/jobs/9016327685?pr=5702#step:9:681) because the new [`shutdown_concurrent_spawn`](https://github.com/tokio-rs/tokio/pull/5702/files#diff-863a5a8b3d134daa64a3f48e856431501050d5dbe3810cd09e8ca10ea96ac73dR1322) test is not ignored for WASM targets, and spawns a thread, which isn't supported in WASM...